Etwas Feinschliff am Event System
Changelog:
- unsichtbares EventSprite hinzugefügt. Damit man die Position der
Events wieder sehen kann, muss man in Settings den Debug Modus
anschalten.

+++ src/Entity/Events/Ambush.java
@@ -10,7 +10,7 @@ public class Ambush extends Event {
 	public Ambush(float x, float y, int eventSize) throws SlickException {
 		super(x, y, eventSize);
 		
-		super.eventChance = randomCalc(1, 10);
+		super.eventChance = (int)randomCalc(1, 10);
 		super.loadImage();
 	}
